requirement.txt
annotated-types==0.5.0
anyio==3.7.1
cached-property==1.5.2
charset-normalizer==3.4.0
colorama==0.4.6
comtypes==1.4.6
cycler==0.11.0
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.2
fonttools==4.38.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib-metadata==6.7.0
kiwisolver==1.4.5
markdown-it-py==2.2.0
matplotlib==3.5.3
mdurl==0.1.2
numpy==1.21.6
openai==1.39.0
opencv-python==4.10.0.84
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pipreqs==0.4.13
pydantic==2.5.3
pydantic_core==2.14.6
pyglossary==4.5.0
Pygments==2.17.2
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pywin32==308
requests==2.31.0
rich==13.8.1
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.7.1
urllib3==2.0.7
wincertstore==0.2
yarg==0.1.10
zipp==3.15.0
